Reformatted code into a style that makes it easier for me to debug and maintain.
This is something I try to avoid, but since since this code appears to have been abandoned and I'm going to maintain this version moving forward, please forgive me for this transgression :).

FWIW, WebStorm informs me that there are 3417 warnings on this commit. None of them originated from the reformat.